drivers: wifi: nrf71: Move firmware interface headers to inc/common/fw_if
Relocate the top-level Wi-Fi firmware interface headers under
inc/common/fw_if, switch consumers to <common/fw_if/...> includes,
and update build, doc, and CODEOWNERS references.
