+++ drivers/nrf71_sr_coex/src/nrf71_sr_coex_cm.c
@@ -26,7 +26,7 @@
 #include <zephyr/logging/log.h>
 
 #include <drivers/wifi/nrf71/nrf71_wifi_coex.h>
-#include <nrf71_coex_if.h>
+#include <common/fw_if/nrf71_coex_if.h>
 
 #include "nrf71_sr_coex_internal.h"
