+++ drivers/wifi/nrf71/inc/common/fw_if/nrf71_cd_sr_if.h
@@ -14,7 +14,7 @@
 #ifndef __NRF71_CD_SR_IF_H__
 #define __NRF71_CD_SR_IF_H__
 
-#include "common/pack_def.h"
+#include <common/fw_if/pack_def.h>
 
 /** Number of elements in CCCONF priority range buffers. */
 #define NUM_ELEMENTS_IN_CCCONF_PTI_RANGE 3U
