+++ drivers/wifi/nrf71/inc/common/fw_if/nrf71_wifi_rf.h
