+++ drivers/wifi/nrf71/inc/common/fw_if/pack_def.h
