+++ drivers/wifi/nrf71/src/common/rf_params.c
@@ -15,7 +15,7 @@
 #include <zephyr/logging/log.h>
 
 #include <common/mem_mgmt.h>
-#include <nrf71_wifi_rf.h>
+#include <common/fw_if/nrf71_wifi_rf.h>
 #include <common/rf_params.h>
 #include <util.h>
